r9 done

+++ sql/r9a.sql
@@ -1,8 +1,11 @@
 /* INSERT PLAYERS INTO WATCHLIST */
 INSERT INTO bookmarks VALUES (..., ...);
 
-/* GET PLAYERS FROM WATCHLIST */
-SELECT * FROM bookmarks JOIN player WHERE player.pid = ...;
+/* GET SINGLE PLAYERS FROM WATCHLIST */
+SELECT * FROM bookmarks, player WHERE player.pid = ... and bookmarks.pid = player.pid;
+
+/* GET ALL PLAYERS FROM WATCHLIST */
+SELECT * FROM bookmarks NATURAL JOIN player WHERE bookmarks.uid = ...;
 
 /* DELETE PLAYERS FROM WATCHLIST */
-DELETE FROM bookmarks WHERE bookmarks.pid = ...;
+DELETE FROM bookmarks WHERE bookmarks.pid = ... and bookmarks.uid = ...;
